Minor improvements to setup scripts

+++ scripts/setupRepo.sh
@@ -22,7 +22,8 @@ OK "Submodules initialized"
 INFO "Configuring Sparse checkout for the submodules"
 cat > "$BASE_DIR/.git/modules/opensmalltalk-vm/info/sparse-checkout" << EOF
 spur64src
-scripts
+scripts/updateSCCSVersions
+scripts/versionInfoPlist
 platforms/nopsys
 platforms/Cross
 third-party/*.spec
@@ -44,7 +45,10 @@ popd > /dev/null
 OK "Sparse checkout configured"
 
 INFO "Checking whether a compilation config exists (and creating a default one if not)"
-cp -n $BASE_DIR/nopsys/compilation.conf.example $BASE_DIR/nopsys/compilation.conf
+if [ ! -f $BASE_DIR/nopsys/compilation.conf ]
+then
+    cp -n $BASE_DIR/nopsys/compilation.conf.example $BASE_DIR/nopsys/compilation.conf
+fi
 
 INFO "Checking for openlibm"
 if [ ! -d "$OPENLIBM_DIR" ]
